+++ src/test/scala/edu/rit/cs/dataunitylab/jsonoid/discovery/schemas/StringSchemaSpec.scala
@@ -234,7 +234,7 @@ class StringSchemaSpec extends UnitSpec {
       Anomaly(
         "$",
         "string length outside histogram range",
-        AnomalyLevel.Warning
+        AnomalyLevel.Info
       )
     )
   }