+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Add namespace to `AnomalyLevel`
 - Unknown properties are not anomalous if `additionalProperties` is true
+- Make histogram bounds violation `Info` level
 
 ### Fixed
 - Properly handle zero values in histogram anomaly checking